# structured-logging: replace KObjs with KObjSlice for logging