#### What type of PR is this?
/kind feature

#### What this PR does / why we need it:
1. This Pull request replaces the `KObjs` with their better performing counter part `KObjSlice`. Since https://github.com/kubernetes/kubernetes/pull/110724 is now merged, We can reap the benefits of the change that went into the `klog` via https://github.com/kubernetes/klog/pull/322

2. As part of the changes done to address #102430, there was a Pull request that got merged (#106978) which made some changes to how the `klog.V(x).Enabled` checks were being made.  This PR also takes second look at some of those changes to see if the `if klogV := klog.V(x); klogV.Enabled()` wrapping is actually required or they can simply be replaced with the `klog.V(x).Info{f|S}`.  Currently this PR only replaces those items that do not have additional computation done for the logging artifacts under the  `if klogV := klog.V(x); klogV.Enabled()` check with their counter part of `klog.V(x)` calls directly.


#### Which issue(s) this PR fixes:
Fixes #110737 

#### Special notes for your reviewer:
##### Benchmark for `getReplicaSetsWithSameController` under `replca_set.go`
Original changes were retained as is for the `if` condition wrapping for better performance
```bash
# With Changes
BenchmarkGetReplicaSetsWithSameController
BenchmarkGetReplicaSetsWithSameController-12    	 3291576	       355.9 ns/op	     184 B/op	       6 allocs/op
BenchmarkGetReplicaSetsWithSameController-12    	 3359475	       360.1 ns/op	     184 B/op	       6 allocs/op
BenchmarkGetReplicaSetsWithSameController-12    	 3335294	       368.1 ns/op	     184 B/op	       6 allocs/op
BenchmarkGetReplicaSetsWithSameController-12    	 3296937	       361.2 ns/op	     184 B/op	       6 allocs/op
BenchmarkGetReplicaSetsWithSameController-12    	 3275090	       389.7 ns/op	     184 B/op	       6 allocs/op

# Without Changes
BenchmarkGetReplicaSetsWithSameController
BenchmarkGetReplicaSetsWithSameController-12    	 5501926	       202.4 ns/op	      48 B/op	       2 allocs/op
BenchmarkGetReplicaSetsWithSameController-12    	 5227614	       200.0 ns/op	      48 B/op	       2 allocs/op
BenchmarkGetReplicaSetsWithSameController-12    	 6073863	       196.5 ns/op	      48 B/op	       2 allocs/op
BenchmarkGetReplicaSetsWithSameController-12    	 6055270	       199.4 ns/op	      48 B/op	       2 allocs/op
BenchmarkGetReplicaSetsWithSameController-12    	 6022762	       195.6 ns/op	      48 B/op	       2 allocs/op
```


#### Does this PR introduce a user-facing change?
```release-note

```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs
[Usage]: https://github.com/kubernetes/community/pull/6719
```

#### TODO

- [ ] Attach Benchmarking results for the changes