@pohly 2022-11-03
/release-note-none

The only difference is some slight performance improvement.